examples/1-hello-world/agent-code/pyproject.toml
[project]
name = "hello-world-agent"
version = "0.1.0"
description = "Minimal Strands baseline demo agent for Bedrock AgentCore"
requires-python = ">=3.12"

dependencies = [
    "bedrock-agentcore>=1.0.7",
    "strands-agents>=1.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
